Fix active hours timezone bucketing
