+++ src/activity-metrics.ts
@@ -13,15 +13,6 @@ export type ActivityStatsResult = {
   stateDurations: Record<string, number>;
 };
 
-export type ActiveHoursCell = {
-  dayOfWeek: number;
-  hour: number;
-  count: number;
-  avgPerWeek: number;
-};
-
-const ACTIVE_HOURS_EVENT_TYPES = new Set(["working", "blocked", "waiting_user"]);
-
 function bucketStart(timeMs: number, granularity: ActivityGranularity): string {
   const bucket = new Date(timeMs);
   if (granularity === "week") {
@@ -166,46 +157,4 @@ export function computeDailyStatus(
   return Array.from(dailyMap.entries())
     .map(([day, durations]) => ({ day, ...durations }))
     .sort((a, b) => String(a.day).localeCompare(String(b.day)));
-}
-
-export function computeActiveHours(
-  rows: Array<Pick<ActivityEventRow, "event_type" | "created_at">>,
-  rangeStart: Date | null,
-  now = new Date()
-): ActiveHoursCell[] {
-  const counts = new Map<string, number>();
-  const rangeStartMs = rangeStart?.getTime() ?? null;
-  let firstTimestamp = Number.POSITIVE_INFINITY;
-  let lastTimestamp = Number.NEGATIVE_INFINITY;
-
-  for (const row of rows) {
-    const timestamp = row.created_at.getTime();
-    if (rangeStartMs !== null && timestamp < rangeStartMs) continue;
-    if (!ACTIVE_HOURS_EVENT_TYPES.has(row.event_type)) continue;
-
-    firstTimestamp = Math.min(firstTimestamp, timestamp);
-    lastTimestamp = Math.max(lastTimestamp, timestamp);
-
-    const key = `${row.created_at.getUTCDay()}:${row.created_at.getUTCHours()}`;
-    counts.set(key, (counts.get(key) ?? 0) + 1);
-  }
-
-  const effectiveStartMs = rangeStartMs ?? (Number.isFinite(firstTimestamp) ? firstTimestamp : now.getTime());
-  const effectiveEndMs = Number.isFinite(lastTimestamp) ? Math.max(lastTimestamp, now.getTime()) : now.getTime();
-  const spanWeeks = Math.max(1, (effectiveEndMs - effectiveStartMs) / (7 * 24 * 60 * 60 * 1000));
-
-  const cells: ActiveHoursCell[] = [];
-  for (let dayOfWeek = 0; dayOfWeek < 7; dayOfWeek += 1) {
-    for (let hour = 0; hour < 24; hour += 1) {
-      const count = counts.get(`${dayOfWeek}:${hour}`) ?? 0;
-      cells.push({
-        dayOfWeek,
-        hour,
-        count,
-        avgPerWeek: Number((count / spanWeeks).toFixed(2)),
-      });
-    }
-  }
-
-  return cells;
 }
